Add a TWAIN wrapper (from the work branch). If SANE is not found on Windows, the Twain wrapper is used.
The ksane widget is only a button to launch the built-in scan dialogs of the twain enabled devices. It makes it possible to use libksane on windows, but the result is not elegant :(

My itch is not strong enough  to do a generic twain interface, so if there is somebody with that kind of  itch, I would be happy to integrate it and look over the API if changes are needed.

OSX: The code uses LoadLibraryA() to loads the "TWAIN_32.DLL", so I'm not sure how it would work on OSX...

CCMAIL: kde-imaging@kde.org

svn path=/trunk/KDE/kdegraphics/libs/libksane/; revision=997486