Apply patch from Cobus Carstens for 'Unhandled corner case'
I'm almost done with moving all the sane read commands to a thread (including sane_start()) and if I get it to work properly, I will commit that to trunk. That change would also mean a change in the logic this patch addresses. (that is the reason I did not apply it immediately to trunk :)

Thanks again for the patch :)

CCBUG:214679

svn path=/trunk/KDE/kdegraphics/libs/libksane/; revision=1049925