Add public API for collapsing the option tabs.
CCMAIL:kde-imaging@kde.org

I had planned to add public API for collapsing the option with the "splitter collapser", but I forgot about it until a couple of days ago. I see this as such a minor change that I would like to add it even tho I have not  added it to  the feature plan for 4.4 before the soft freeze.

I will revert the change if i get objections.

svn path=/trunk/KDE/kdegraphics/libs/libksane/; revision=1053346