Fix translations from Sane
Summary:
The strings from sane requires a different translation domain.
Wraps the i18n calls to a special function instead of simply
redefining TRANSLATION_DOMAIN because in some files the calls
to *i18n* with the default domain are interleaved with the
i18n calls to strings from Sane.
Remove the old (and now useless) commented porting notes about
re-adding the support for the sane-backends translation domain.

Test Plan:
The strings are shown when KSaneWidget is used (for example
inside Skanlite).

Reviewers: sars

Reviewed By: sars

Subscribers: #localization

Differential Revision: https://phabricator.kde.org/D9883