Fix setting gamma-tables that are not symmetrical (items / range)
Libksane expected gamma-tables to have as many entries as the maximum
value in the table. This is not the case for example in the niash
backend. This patch will hopefully fix the setting of gamma-table
for the niash backend.

I do not have a scanner that has an asymmetrical gamma-table, so I
cannot test this fully.

BUG: 370972
FIXED-IN: KDE Applications 18.12

Also change names of some variables to follow KF5 coding standards