Add scan-area sliders and predefined page sizes combobox
Summary: Add scan-area sliders and a combo-box that contains predefined page sizes, but only the ones that fit in the maximum scan area of the scanner. The combo-box also contains landscape versions of the sizes that fit.

Test Plan:
The sliders and combo-box interact with the normal selection and does not break it.

{F7856101}

Reviewers: antonarnold, #vdg, #kde_applications, ngraham

Reviewed By: antonarnold, #vdg, ngraham

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D26351