KSaneOptCombo: refactor handling of string values
We need to handle the string values as untranslated when getting/setting
them, otherwise it breaks when saving them to file, or checking for
a specific combobox value.

Hence, refactor genComboStringList() so it returns a list of pairs of
<identifier, translated string>, adding the translated strings as texts
in the combobox, with the identifiers as item data. Then, use the item
data to get/set the value we want.